♻️: Move `App.tsx` into `src` directory
`src` seems to be used more widely than `app`
