+++ .github/workflows/typescript.yaml
@@ -9,7 +9,7 @@ on:
       - '**/*.jsx'
       - '**/*.ts'
       - '**/*.tsx'
-      - 'app/**'
+      - 'src/**'
       - '__tests__/**'
       - '__mocks__/**'
       - 'app.json'
@@ -29,7 +29,7 @@ on:
       - '**/*.jsx'
       - '**/*.ts'
       - '**/*.tsx'
-      - 'app/**'
+      - 'src/**'
       - '__tests__/**'
       - '__mocks__/**'
       - 'app.json'
