+++ README.md
@@ -100,3 +100,4 @@ npm run -s lint:tsc
 
 - [x] Editorconfig, ESLint, Prettier を追加
 - [x] TypeScript の設定ファイルを修正、`tsc`での型チェックを lint に追加
+- [x] `App.tsx`を`app`ディレクトリ配下に移動
