+++ __tests__/App.js
@@ -2,7 +2,7 @@ import 'react-native';
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import App from '../App.tsx';
+import App from '../src/App.tsx';
 
 // Note: test renderer must be required after react-native.
